Update TradeIndicators.xml
What Are Fibonacci Retracement Levels?
Fibonacci retracement levels are horizontal lines that indicate where support and resistance are likely to occur. They are based on Fibonacci numbers. Each level is associated with a percentage. The percentage is how much of a prior move the price has retraced. The Fibonacci retracement levels are 23.6%, 38.2%, 61.8%, and 78.6%. While not officially a Fibonacci ratio, 50% is also used.

Fibonacci Length changed 90 bars to 144 bars.

RSI Signal is added to Indicator.

What is RSI?
The relative strength index (RSI) is a momentum indicator used in technical analysis that measures the magnitude of recent price changes to evaluate overbought or oversold conditions in the price of a stock or other asset. The RSI is displayed as an oscillator (a line graph that moves between two extremes) and can have a reading from 0 to 100. The indicator was originally developed by J. Welles Wilder Jr. and introduced in his seminal 1978 book, "New Concepts in Technical Trading Systems."

EMA Signal is added to Indicator.

What is EMA?
An exponential moving average (EMA) is a type of moving average (MA) that places a greater weight and significance on the most recent data points. The exponential moving average is also referred to as the exponentially weighted moving average.

Timestamp input is added to indicator.

EMA Signal is removed from indicator.

SMA 50 and MA 200 is added to indicator.

What is MA?
A simple moving average (SMA) calculates the average of a selected range of prices, usually closing prices, by the number of periods in that range.

Golden Cross, Silver Cross and Death Cross signals are added to indicator.

What is Golden, Silver, Death Cross?
Golden Cross; occurs when a short-term moving average crosses over a major long-term moving average to the upside and is interpreted by analysts and traders as signaling a definitive upward turn in a market. Basically, the short-term average trends up faster than the long-term average, until they cross.

Death Cross; Conversely, a similar downside moving average crossover constitutes the death cross and is understood to signal a decisive downturn in a market. The death cross occurs when the short term average trends down and crosses the long-term average, basically going in the opposite direction of the golden cross.

Silver Cross; Silver Cross; It is one of the technical analysis-based terms that signals the rise in the stock market. It occurs when the rise seen in the 21-day moving averages cuts the 50-day moving average.

Timestamp condition is added to indicator.

SMA custom inputs has been added to indicator.

Bollinger Band Buy and Sell signals has been added to the indicator.

Several improvements on code.

What is Bollinger Band : A Bollinger Band is a technical analysis tool defined by a set of trendlines plotted two standard deviations (positively and negatively) away from a simple moving average (SMA) of a security's price, but which can be adjusted to user preferences. Bollinger Bands were developed and copyrighted by famous technical trader John Bollinger, designed to discover opportunities that give investors a higher probability of properly identifying when an asset is oversold or overbought

Several improvements on code.

What is Williams %R? - Williams %R, also known as the Williams Percent Range, is a type of momentum indicator that moves between 0 and -100 and measures overbought and oversold levels. The Williams %R may be used to find entry and exit points in the market. The indicator is very similar to the Stochastic oscillator and is used in the same way. It was developed by Larry Williams and it compares a stock’s closing price to the high-low range over a specific period, typically 14 days or periods.